+++ R/FINEMAP_check_version.R
@@ -1,4 +1,5 @@
-FINEMAP_check_version <- function(FINEMAP_path,
+FINEMAP_check_version <- function(FINEMAP_path=
+                                      FINEMAP_find_executable(verbose = FALSE),
                                   verbose=TRUE){
     #### FINEMAP does not have a -v or --version flag. ####
     out <- system(paste(FINEMAP_path,"-h"), intern = TRUE)