+++ R/POLYFUN_help.R
@@ -4,21 +4,24 @@
 #' Will automatically install PolyFun if has not been installed already.
 #' @param polyfun [Optional] Path to PolyFun executable.
 #' @inheritParams echoconda::find_python_path
-#' @export
 #' @family polyfun
 #' @returns PolyFun help message as character vector.
+#' 
+#' @export
+#' @importFrom echoconda find_python_path
 #' @examples 
-#' echofinemap::POLYFUN_help()
+#' h <- echofinemap::POLYFUN_help()
 POLYFUN_help <- function(polyfun=NULL,
-                         conda_env="echoR_mini"){
-    POLYFUN_install()
-    polyfun <- POLYFUN_find_polyfun_folder(polyfun_path = polyfun)
+                         conda_env="echoR_mini",
+                         verbose=TRUE){
+    POLYFUN_install(verbose = verbose)
+    polyfun <- POLYFUN_find_folder(polyfun_path = polyfun)
     python <- echoconda::find_python_path(conda_env = conda_env)
-    messager("Retrieving PolyFun --help documentation.")
+    messager("Retrieving PolyFun --help documentation.",v=verbose)
     cmd <- paste(python,
                  file.path(polyfun,"polyfun.py"),
                  "--help")
     out <- system(cmd, intern = TRUE)
-    message(paste(out,collapse = "\n"))
+    messager(paste(out,collapse = "\n"), v=verbose)
     return(out)
 }