+++ R/POLYFUN_install.R
@@ -4,8 +4,10 @@
 #' @export
 #' @param force_new Force a new installation 
 #' even when PolyFun is already installed. 
+#' @param verbose Print messages.
 #' @returns Null.
-POLYFUN_install <- function(force_new=FALSE){
+POLYFUN_install <- function(force_new=FALSE,
+                            verbose=TRUE){
    POLYFUN_installed <- function(){
        dirs <- grep("polyfun", 
                     list.dirs(system.file("tools",package = "echofinemap"),
@@ -20,12 +22,12 @@ POLYFUN_install <- function(force_new=FALSE){
                            "This will automatically restart R. (y/n): ")
         )
         if(tolower(install)=="y"){
-            messager("Installing PolyFun submodule.")
+            messager("Installing PolyFun submodule.",v=verbose)
             install_submodule_git("https://github.com/RajLabMSSM/echofinemap")
         }
         is_installed <- POLYFUN_installed()
         if(isFALSE(is_installed)) stop("PolyFun is not installed.")
     } else {
-        messager("PolyFun submodule already installed.")
+        messager("PolyFun submodule already installed.",v=verbose)
     }
 }